# BackEnd/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "ITSC .NET Practice API",
    "description": "API documentation for ITSC .NET Practice project",
    "version": "v1"
  },
  "paths": {
    "/api/admin/login": {
      "get": {
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Dtos.LoginRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Dtos.LoginRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Dtos.LoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Dtos.RegisterRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Dtos.RegisterRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Dtos.RegisterRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/orders": {
      "get": {
        "tags": [
          "Order"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Order"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Dtos.OrderDto.OrderRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Dtos.OrderDto.OrderRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Dtos.OrderDto.OrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/status/{status}": {
      "get": {
        "tags": [
          "Order"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/user/{userId}": {
      "get": {
        "tags": [
          "Order"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/user/{userId}/status/{status}": {
      "get": {
        "tags": [
          "Order"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "status",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{orderId}": {
      "put": {
        "tags": [
          "Order"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{orderId}/cancel": {
      "delete": {
        "tags": [
          "Order"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
                }
              }
            }
          }
        }
      }
    },
    "/api/products": {
      "get": {
        "tags": [
          "Product"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Dtos.ProductDto.Request"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Dtos.ProductDto.Request"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Dtos.ProductDto.Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Product"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Product"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Product"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/products/search": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "itsc_dotnet_practice.Models.Dtos.LoginRequestDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "itsc_dotnet_practice.Models.Dtos.OrderDetailDto.OrderDetailRequest": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "itsc_dotnet_practice.Models.Dtos.OrderDto.OrderRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "orderDetails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Dtos.OrderDetailDto.OrderDetailRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "itsc_dotnet_practice.Models.Dtos.ProductDto.Request": {
        "required": [
          "description",
          "name",
          "price",
          "stock"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "minLength": 1,
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "itsc_dotnet_practice.Models.Dtos.RegisterRequestDto": {
        "required": [
          "confirmPassword",
          "fullName",
          "password",
          "phone",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "minLength": 1,
            "type": "string"
          },
          "fullName": {
            "minLength": 1,
            "type": "string"
          },
          "phone": {
            "minLength": 1,
            "type": "string",
            "format": "tel"
          },
          "password": {
            "minLength": 6,
            "type": "string"
          },
          "confirmPassword": {
            "minLength": 1,
            "type": "string"
          },
          "role": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "itsc_dotnet_practice.Models.Order": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "minLength": 1,
            "type": "string"
          },
          "shippingAddress": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/itsc_dotnet_practice.Models.User"
          },
          "orderDetails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/itsc_dotnet_practice.Models.OrderDetail"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "itsc_dotnet_practice.Models.OrderDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "productName": {
            "type": "string",
            "nullable": true
          },
          "productImageUrl": {
            "type": "string",
            "nullable": true
          },
          "productDescription": {
            "type": "string",
            "nullable": true
          },
          "productCategory": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "order": {
            "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Order"
          },
          "orderId": {
            "type": "integer",
            "format": "int32"
          },
          "product": {
            "$ref": "#/components/schemas/itsc_dotnet_practice.Models.Product"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "user": {
            "$ref": "#/components/schemas/itsc_dotnet_practice.Models.User"
          }
        },
        "additionalProperties": false
      },
      "itsc_dotnet_practice.Models.Product": {
        "required": [
          "description",
          "name",
          "price",
          "stock"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "minLength": 1,
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "itsc_dotnet_practice.Models.User": {
        "required": [
          "fullName",
          "password",
          "phone",
          "username"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "minLength": 1,
            "type": "string"
          },
          "fullName": {
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "minLength": 1,
            "type": "string"
          },
          "phone": {
            "minLength": 1,
            "type": "string"
          },
          "role": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "description": "Basic Authentication",
        "scheme": "basic"
      },
      "BearerAuth": {
        "type": "http",
        "description": "JWT Bearer Authentication",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BasicAuth": [ ],
      "BearerAuth": [ ]
    }
  ]
}